Completed the functions for the app

+++ resources/js/components/Login.vue
@@ -1,17 +1,40 @@
 <template>
-    <div>
-        <h2>Login</h2>
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="col-md-6">
+                <div class="card">
+                    <div class="card-header">Login</div>
 
-        <form @submit.prevent="login" class="mb-3">
-            <div class="form-group">
-                <input type="text" class="form-control" placeholder="Email" v-model="user.email">
+                    <div class="card-body">
+                        <form @submit.prevent="login" class="mb-3">
+                            <div class="form-group">
+                                <input
+                                    type="text"
+                                    class="form-control"
+                                    placeholder="Email"
+                                    v-model="user.email"
+                                />
+                            </div>
+                            <div class="form-group">
+                                <input
+                                    type="password"
+                                    class="form-control"
+                                    placeholder="Password"
+                                    v-model="user.password"
+                                />
+                            </div>
+
+                            <button
+                                type="submit"
+                                class="btn btn-light btn-block"
+                            >
+                                Login
+                            </button>
+                        </form>
+                    </div>
+                </div>
             </div>
-            <div class="form-group">
-                <input type="password" class="form-control" placeholder="Password" v-model="user.password">
-            </div>
-            
-            <button type="submit" class="btn btn-light btn-block">Login</button>
-        </form>
+        </div>
     </div>
 </template>
 
@@ -20,30 +43,40 @@ export default {
     data() {
         return {
             user: {
-                email: '',
-                password: ''
+                email: "",
+                password: ""
             }
         };
     },
 
+    created() {
+        if (localStorage.getItem("user")) {
+            this.$router.push("./to-do-lists");
+        }
+    },
+
     methods: {
-        login(){
-            const headers = { 
+        login() {
+            const headers = {
                 "Content-Type": "application/json",
-                "Accept": "application/json"
+                Accept: "application/json"
             };
 
-            fetch('api/login', { 
-                    method: 'post',
-                    headers: headers,
-                    body: JSON.stringify(this.user)
-                })
+            fetch("api/login", {
+                method: "post",
+                headers: headers,
+                body: JSON.stringify(this.user)
+            })
                 .then(res => res.json())
                 .then(res => {
-                    localStorage.clear();
-                    localStorage.setItem('user', JSON.stringify(res.data));
-
-                    this.$router.push('./to-do-lists');
+                    if (res.success) {
+                        localStorage.clear();
+                        localStorage.setItem("user", JSON.stringify(res.data));
+                        
+                        this.$router.push("./to-do-lists");
+                    }else {
+                        alert('Incorrect email or password');
+                    }
                 })
                 .catch(err => console.log(err));
         }
